cfs/pyproject.toml
[project]
name = "cfs-local"
version = "0.1.0"
description = "Computational File System daemon and SDK for KVBFS"
requires-python = ">=3.10"
dependencies = [
    "llama-cpp-python>=0.2.90",
    "inotify-simple>=1.3.5",
    "pydantic>=2.0",
]

[project.optional-dependencies]
async = ["aiofiles>=23.0"]
dev = ["pytest>=7.0"]
